+++ paper.bib
@@ -97,6 +97,8 @@ available from http://www.sbml.org/Contact: sysbio-team@caltech.edu*To whom corr
   volume={20},
   number={13},
   pages={2143--2144},
+  url = {https://doi.org/10.1093/bioinformatics/bth200},
+  doi = {10.1093/bioinformatics/bth200},
   year={2004},
   publisher={Oxford University Press}
 }
