### 2023-04-14 b424aa3
- add support for advanced event flags
