### 2023-04-14 82729cf
- if there are no fit_parameters don't try to set them.
