### 2023-04-17 536524f
- add petab test
some of the test require environment variables:

PETAB_BENCHMARK_MODELS

and / or

PETAB_SELECT_MODELS

defined with paths to the models
