+++ tests/model_selection/petab_select_problem.yaml
@@ -1,0 +1,5 @@
+version: beta_1
+criterion: AIC
+method: forward
+model_space_files:
+- model_space.tsv
