### 2023-04-17 d1e4381
- add petab test
some of the test require environment variables:

PETAB_BENCHMARK_MODELS

and / or

PETAB_SELECT_MODELS

defined with paths to the models
