+++ tests/model_selection/calibrated/M1_4_0.yaml
@@ -1,0 +1,11 @@
+criteria:
+  AIC: 30
+model_id: M1_4_0
+parameters:
+  k1: 0.2
+  k2: estimate
+  k3: estimate
+estimated_parameters:
+  k2: 0.15
+  k3: 0.0
+petab_yaml: ../petab_problem.yaml
