+++ docs/notebooks/temp_selection/cps_c93a94894cd374093ca975a65488754aae48acd355cf387059878eaa3f956fb967ed3014c2087fcda647c616e076754d37582e53444e6bced08272ab4a2a5c87_report.txt
@@ -1,87 +1,0 @@
-Parameter Estimation Task
-
-Problem Description:
-Subtask: 
-Time-Course Task
-
-Problem Description:
-    AutomaticStepSize: 0
-    StepNumber: 100
-    StepSize: 0.01
-    Duration: 1
-    TimeSeriesRequested: 1
-    OutputStartTime: 0
-    Output Event: 0
-    Start in Steady State: 0
-    Use Values: 0
-    Values: 
-
-Method: Deterministic (LSODA)
-    Integrate Reduced Model: 0
-    Relative Tolerance: 1e-06
-    Absolute Tolerance: 1e-12
-    Max Internal Steps: 100000
-    Max Internal Step Size: 0
-
-
-List of Fitting Items:
-    0.0 <= Values[k1].InitialValue <= 1000.0; Start Value = 0.201609
-    Affected Experiments:
-      all    Affected Validation Experiments:
-      all
-    0.0 <= Values[k3].InitialValue <= 1000.0; Start Value = 0
-    Affected Experiments:
-      all    Affected Validation Experiments:
-      all
-    1e-05 <= Values[sigma_x2].InitialValue <= 1000.0; Start Value = 0.162347
-    Affected Experiments:
-      all    Affected Validation Experiments:
-      all
-
-List of Constraint Items:
-
-Method: Levenberg - Marquardt
-    Log Verbosity: 0
-    Iteration Limit: 2000
-    Tolerance: 1e-06
-    Modulation: 1e-06
-    Stop after # stalled iterations: 0
-    Initial Lambda: 1
-    Lambda Decrease: 2
-    Lambda Increase: 4
-
-[Function Evaluations]	[Best Value]	[Best Parameters]
-1	0.0823312	(	0.201609	0	0.162347	)
-
-Parameter Estimation Result:
-
-Objective Function Value:	0.0823312
-Standard Deviation:	0.165661
-Function Evaluations:	140
-CPU Time [s]:	0.015
-Evaluations/Second [1/s]:	9333.33
-
-	Parameter	Value	Gradient	Standard Deviation
-	Values[k1].InitialValue:	0.201609	nan	nan
-	Values[k3].InitialValue:	0	nan	nan
-	Values[sigma_x2].InitialValue:	0.162347	nan	nan
-
-File Name:	C:/Development/basico/docs/notebooks/temp_selection/cps_c93a94894cd374093ca975a65488754aae48acd355cf387059878eaa3f956fb967ed3014c2087fcda647c616e076754d37582e53444e6bced08272ab4a2a5c87.txt
-Experiment:	model1_data1
-Mean:	0.021549
-Objective Value:	0.0823312
-Root Mean Square:	0.11714
-Row	Time	Values[obs_x2](Data)	Values[obs_x2](Fit)	Values[obs_x2](Weighted Error)	Objective Value	Root Mean Square
-
-1.	0	0	0	0	0	0
-2.	1	0.194218	0.00975266	-0.184465	0.0340273	0.184465
-3.	5	0.0484032	0.214775	0.166372	0.0276796	0.166372
-4.	10	0.61288	0.741677	0.128797	0.0165887	0.128797
-5.	30	4.07931	4.13255	0.0532423	0.00283474	0.0532423
-6.	60	10.1201	10.0854	-0.0346524	0.00120079	0.0346524
-Objective Value				0.0823312
-Root Mean Square				0.11714
-Weight				1
-
-
-
