+++ docs/notebooks/temp_selection/cps_d234d7aa67f86f33949da8bdbd499da46852a197e33219cd6d1763009eca13e243de9c049e80442c0fd718ca9fd80bd9d0c2666265ffc9b29ebbfcdd53f22336_report.txt
@@ -1,206 +1,0 @@
-Parameter Estimation Task
-
-Problem Description:
-Subtask: 
-Time-Course Task
-
-Problem Description:
-    AutomaticStepSize: 0
-    StepNumber: 100
-    StepSize: 0.01
-    Duration: 1
-    TimeSeriesRequested: 1
-    OutputStartTime: 0
-    Output Event: 0
-    Start in Steady State: 0
-    Use Values: 0
-    Values: 
-
-Method: Deterministic (LSODA)
-    Integrate Reduced Model: 0
-    Relative Tolerance: 1e-06
-    Absolute Tolerance: 1e-12
-    Max Internal Steps: 100000
-    Max Internal Step Size: 0
-
-
-List of Fitting Items:
-    0.0 <= Values[k1].InitialValue <= 1000.0; Start Value = 0.268092
-    Affected Experiments:
-      all    Affected Validation Experiments:
-      all
-    0.0 <= Values[k2].InitialValue <= 1000.0; Start Value = 0.058504
-    Affected Experiments:
-      all    Affected Validation Experiments:
-      all
-    0.0 <= Values[k3].InitialValue <= 1000.0; Start Value = 0.00100447
-    Affected Experiments:
-      all    Affected Validation Experiments:
-      all
-    1e-05 <= Values[sigma_x2].InitialValue <= 1000.0; Start Value = 0.224285
-    Affected Experiments:
-      all    Affected Validation Experiments:
-      all
-
-List of Constraint Items:
-
-Method: Levenberg - Marquardt
-    Log Verbosity: 0
-    Iteration Limit: 2000
-    Tolerance: 1e-06
-    Modulation: 1e-06
-    Stop after # stalled iterations: 0
-    Initial Lambda: 1
-    Lambda Decrease: 2
-    Lambda Increase: 4
-
-[Function Evaluations]	[Best Value]	[Best Parameters]
-1	0.0532241	(	0.268092	0.058504	0.00100447	0.224285	)
-16	0.0531987	(	0.268093	0.0585044	0.000995648	0.224285	)
-22	0.0531761	(	0.268093	0.0585042	0.000998686	0.224285	)
-27	0.0531758	(	0.268093	0.0585042	0.000999627	0.224285	)
-33	0.0531756	(	0.268093	0.0585042	0.000999305	0.224285	)
-38	0.0531756	(	0.268093	0.0585041	0.000999212	0.224285	)
-44	0.0531756	(	0.268093	0.0585041	0.000999243	0.224285	)
-49	0.0531756	(	0.268093	0.0585041	0.000999248	0.224285	)
-54	0.0531755	(	0.268093	0.058504	0.000999232	0.224285	)
-59	0.0531755	(	0.268093	0.0585039	0.000999308	0.224285	)
-66	0.0531755	(	0.268093	0.0585039	0.00099925	0.224285	)
-71	0.0531755	(	0.268093	0.0585039	0.000999227	0.224285	)
-76	0.0531755	(	0.268093	0.0585038	0.00099927	0.224285	)
-82	0.0531755	(	0.268093	0.0585038	0.000999222	0.224285	)
-87	0.0531755	(	0.268093	0.0585037	0.000999261	0.224285	)
-93	0.0531755	(	0.268093	0.0585037	0.000999217	0.224285	)
-98	0.0531755	(	0.268093	0.0585036	0.000999269	0.224285	)
[…]

[…]
