+++ docs/notebooks/temp_selection/cps_d234d7aa67f86f33949da8bdbd499da46852a197e33219cd6d1763009eca13e243de9c049e80442c0fd718ca9fd80bd9d0c2666265ffc9b29ebbfcdd53f22336.txt
@@ -1,7 +1,0 @@
-#time	obs_x2
-0	0.0
-1	0.19421762
-5	0.0484032
-10	0.61288016
-30	4.07930835
-60	10.12008893
