+++ docs/notebooks/temp_selection/cps_5274a379cce41a92e8bf549bbf3245ea83e273fc59837660dc861281577a4b19b276663dfcb85aa267bc6aa183fd95509f5d4ea42e2e5b898da2465e49cff925_report.txt
@@ -1,87 +1,0 @@
-Parameter Estimation Task
-
-Problem Description:
-Subtask: 
-Time-Course Task
-
-Problem Description:
-    AutomaticStepSize: 0
-    StepNumber: 100
-    StepSize: 0.01
-    Duration: 1
-    TimeSeriesRequested: 1
-    OutputStartTime: 0
-    Output Event: 0
-    Start in Steady State: 0
-    Use Values: 0
-    Values: 
-
-Method: Deterministic (LSODA)
-    Integrate Reduced Model: 0
-    Relative Tolerance: 1e-06
-    Absolute Tolerance: 1e-12
-    Max Internal Steps: 100000
-    Max Internal Step Size: 0
-
-
-List of Fitting Items:
-    0.0 <= Values[k2].InitialValue <= 1000.0; Start Value = 0.101478
-    Affected Experiments:
-      all    Affected Validation Experiments:
-      all
-    0.0 <= Values[k3].InitialValue <= 1000.0; Start Value = 0
-    Affected Experiments:
-      all    Affected Validation Experiments:
-      all
-    1e-05 <= Values[sigma_x2].InitialValue <= 1000.0; Start Value = 0.149996
-    Affected Experiments:
-      all    Affected Validation Experiments:
-      all
-
-List of Constraint Items:
-
-Method: Levenberg - Marquardt
-    Log Verbosity: 0
-    Iteration Limit: 2000
-    Tolerance: 1e-06
-    Modulation: 1e-06
-    Stop after # stalled iterations: 0
-    Initial Lambda: 1
-    Lambda Decrease: 2
-    Lambda Increase: 4
-
-[Function Evaluations]	[Best Value]	[Best Parameters]
-1	0.0884601	(	0.101478	0	0.149996	)
-
-Parameter Estimation Result:
-
-Objective Function Value:	0.0884601
-Standard Deviation:	0.171717
-Function Evaluations:	140
-CPU Time [s]:	0.000
-Evaluations/Second [1/s]:	inf
-
-	Parameter	Value	Gradient	Standard Deviation
-	Values[k2].InitialValue:	0.101478	nan	nan
-	Values[k3].InitialValue:	0	nan	nan
-	Values[sigma_x2].InitialValue:	0.149996	nan	nan
-
-File Name:	C:/Development/basico/docs/notebooks/temp_selection/cps_5274a379cce41a92e8bf549bbf3245ea83e273fc59837660dc861281577a4b19b276663dfcb85aa267bc6aa183fd95509f5d4ea42e2e5b898da2465e49cff925.txt
-Experiment:	model1_data1
-Mean:	0.0117971
-Objective Value:	0.0884601
-Root Mean Square:	0.121422
-Row	Time	Values[obs_x2](Data)	Values[obs_x2](Fit)	Values[obs_x2](Weighted Error)	Objective Value	Root Mean Square
-
-1.	0	0	0	0	0	0
-2.	1	0.194218	0.00981311	-0.184405	0.034005	0.184405
-3.	5	0.0484032	0.215722	0.167319	0.0279956	0.167319
-4.	10	0.61288	0.743537	0.130656	0.0170711	0.130656
-5.	30	4.07931	4.123	0.0436941	0.00190918	0.0436941
-6.	60	10.1201	10.0336	-0.0864825	0.00747922	0.0864825
-Objective Value				0.0884601
-Root Mean Square				0.121422
-Weight				1
-
-
-
