+++ docs/notebooks/temp_selection/problem_12a5825db2d08e1c71e0c87771592dd9adc9b550beb389cad8c1cce4ed3fbaa6a7ed484462687bd85d61c221efc185224f2b612f8f6867e6c4c0930d55358be7.yaml
@@ -1,11 +1,0 @@
-parameter_file: parameters_12a5825db2d08e1c71e0c87771592dd9adc9b550beb389cad8c1cce4ed3fbaa6a7ed484462687bd85d61c221efc185224f2b612f8f6867e6c4c0930d55358be7.tsv
-format_version: 1
-problems:
-- condition_files:
-  - experimentalCondition_12a5825db2d08e1c71e0c87771592dd9adc9b550beb389cad8c1cce4ed3fbaa6a7ed484462687bd85d61c221efc185224f2b612f8f6867e6c4c0930d55358be7.tsv
-  measurement_files:
-  - measurementData_12a5825db2d08e1c71e0c87771592dd9adc9b550beb389cad8c1cce4ed3fbaa6a7ed484462687bd85d61c221efc185224f2b612f8f6867e6c4c0930d55358be7.tsv
-  sbml_files:
-  - model_12a5825db2d08e1c71e0c87771592dd9adc9b550beb389cad8c1cce4ed3fbaa6a7ed484462687bd85d61c221efc185224f2b612f8f6867e6c4c0930d55358be7.xml
-  observable_files:
-  - observables_12a5825db2d08e1c71e0c87771592dd9adc9b550beb389cad8c1cce4ed3fbaa6a7ed484462687bd85d61c221efc185224f2b612f8f6867e6c4c0930d55358be7.tsv

[…]
