### 2023-04-21 15a9455
fix an issue with updating simulation dataframe when times repeat
