### 2023-04-21 883f729
fix an issue with updating simulation dataframe when times repeat
