+++ basico/model_info.py
@@ -28,6 +28,8 @@ import dateutil.parser
 import datetime
 import sys
 
+logger = logging.getLogger(__name__)
+
 MIRIAM_XML = 'http://copasi.org/static/miriam.xml' # noqa
 
 try:
@@ -589,7 +591,7 @@ def get_report_dict(report, **kwargs):
     if isinstance(report, str) or isinstance(report, int):
         spec = dm.getReportDefinition(report)
         if spec is None:
-            logging.error('No plot specification: {0}'.format(report))
+            logger.error('No plot specification: {0}'.format(report))
             return
 
         report = spec
@@ -626,7 +628,7 @@ def _add_report_items_to_list(result, num_elements, get_nth_function, dm):
             obj = dm.getObjectFromCN(COPASI.CCommonName(cn))
 
         if obj is None:
-            logging.warning('report item cannot be resolved: {0}'.format(cn.getString()))
+            logger.warning('report item cannot be resolved: {0}'.format(cn.getString()))
             continue
 
         reverse = dm.findObjectByDisplayName(obj.getObjectDisplayName())
@@ -685,7 +687,7 @@ def get_plot_dict(plot_spec, **kwargs):
     if isinstance(plot_spec, str) or isinstance(plot_spec, int):
         spec = dm.getPlotSpecification(plot_spec)
         if spec is None:
-            logging.error('No plot specification: {0}'.format(plot_spec))
+            logger.error('No plot specification: {0}'.format(plot_spec))
             return
 
         plot_spec = spec
@@ -795,7 +797,7 @@ def set_plot_curves(plot_spec, curves, **kwargs):
     if isinstance(plot_spec, str) or isinstance(plot_spec, int):
         spec = dm.getPlotSpecification(plot_spec)
         if spec is None:
-            logging.error('No plot specification: {0}'.format(plot_spec))
+            logger.error('No plot specification: {0}'.format(plot_spec))
             return
 
         plot_spec = spec
@@ -856,7 +858,7 @@ def set_plot_curves(plot_spec, curves, **kwargs):
                 if channel == 'Time':
                     obj = dm.getModel().getValueReference()
                 if obj is None:
-                    logging.warning("Couldn't resolve {0} when adding curve spec".format(channel))
+                    logger.warning("Couldn't resolve {0} when adding curve spec".format(channel))
                     continue
                 plot_item.addChannel(COPASI.CPlotDataChannelSpec(obj.getCN()))
 
@@ -907,7 +909,7 @@ def add_default_plot(name, **kwargs):
     """
 
     if name not in get_default_plot_names():
-        logging.warning('No such default plot: {0}'.format(name))
+        logger.warning('No such default plot: {0}'.format(name))
         return
 
     item = COPASI.COutputAssistant.getItem(_default_plots[name])
@@ -925,11 +927,11 @@ def add_default_plot(name, **kwargs):
             task.initializeRaw(COPASI.CCopasiTask.OUTPUT_UI)
             result = COPASI.COutputAssistant.createDefaultOutput(_default_plots[name], task, dm)
             if result is None:
-                logging.warning('Failed to create default plot for: {0}'.format(name))
+                logger.warning('Failed to create default plot for: {0}'.format(name))
                 return None
             return result.getObjectName()
 
-    logging.warning('No task found for the plot')
+    logger.warning('No task found for the plot')
     return None
 
 
@@ -996,7 +998,7 @@ def set_plot_dict(plot_spec, active=True, log_x=False, log_y=False, tasks='', **
     if isinstance(plot_spec, str) or isinstance(plot_spec, int):
         spec = dm.getPlotSpecification(plot_spec)
         if spec is None:
[…]
