### 2023-05-16 093b9c6
- change default evaluation strategy
