### 2023-05-19 #5325 @Robaina
Hi @marcosvital,

Apologies, I'm already reviewing a package at pyOpenSci at the moment. I would suggest `ayush9pandey`, he submitted a package to JOSS not long ago about a similar topic (I was a reviewer).

Cheers
