### 2023-05-26 ff8774b
- add function to transform results back to nominal values
