### 2023-05-26 449c0b7
- add function to transform results back to nominal values
