### 2023-06-09 da63f74
#33 : add missing simulation method
