### 2023-06-13 #5553 @editorialbot
[REVIEW]: BASICO: A simplified python interface to COPASI
**Submitting author:** @fbergmann (<a href="http://orcid.org/0000-0001-5553-4702">Frank T. Bergmann</a>)
**Repository:** https://github.com/copasi/basico
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v0.52
