### 2023-06-16 1fb8692
- avoid None references
