+++ docs/API/basico.rst
@@ -9,18 +9,11 @@ basico package
 Submodules
 ----------
 
-basico.array\_tools module
---------------------------
-
-.. automodule:: basico.array_tools
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 basico.biomodels module
 -----------------------
 
 .. automodule:: basico.biomodels
+   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
@@ -29,6 +22,7 @@ basico.compartment\_array\_tools module
 ---------------------------------------
 
 .. automodule:: basico.compartment_array_tools
+   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
@@ -37,6 +31,7 @@ basico.jws\_online module
 -------------------------
 
 .. automodule:: basico.jws_online
+   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
@@ -45,6 +40,7 @@ basico.model\_info module
 -------------------------
 
 .. automodule:: basico.model_info
+   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
@@ -53,6 +49,7 @@ basico.model\_io module
 -----------------------
 
 .. automodule:: basico.model_io
+   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
@@ -61,6 +58,7 @@ basico.task\_parameterestimation module
 ---------------------------------------
 
 .. automodule:: basico.task_parameterestimation
+   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
@@ -69,6 +67,7 @@ basico.task\_steadystate module
 -------------------------------
 
 .. automodule:: basico.task_steadystate
+   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
@@ -77,6 +76,7 @@ basico.task\_timecourse module
 ------------------------------
 
 .. automodule:: basico.task_timecourse
+   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
@@ -85,6 +85,7 @@ basico.task\_scan module
 ------------------------------
 
 .. automodule:: basico.task_scan
+   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
@@ -94,6 +95,16 @@ basico.task\_optimization module
 ------------------------------
 
 .. automodule:: basico.task_optimization
+   :imported-members:
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+basico.task\_sensitivities module
+------------------------------
+
+.. automodule:: basico.task_sensitivities
+   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
