### 2023-06-16 d71d07c
- fix required packages for doc generation
