+++ docs/requirements.txt
@@ -5,7 +5,9 @@ numpy
 pandoc
 python-copasi
 cython
+PyYAML
 pandas
+scipy
 sphinx>3
 sphinx_rtd_theme
 pygments>=2.4.1
