+++ docs/API/basico.rst
@@ -13,7 +13,6 @@ basico.biomodels module
 -----------------------
 
 .. automodule:: basico.biomodels
-   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
@@ -22,7 +21,6 @@ basico.compartment\_array\_tools module
 ---------------------------------------
 
 .. automodule:: basico.compartment_array_tools
-   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
@@ -31,7 +29,6 @@ basico.jws\_online module
 -------------------------
 
 .. automodule:: basico.jws_online
-   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
@@ -40,7 +37,6 @@ basico.model\_info module
 -------------------------
 
 .. automodule:: basico.model_info
-   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
@@ -49,7 +45,6 @@ basico.model\_io module
 -----------------------
 
 .. automodule:: basico.model_io
-   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
@@ -58,7 +53,6 @@ basico.task\_parameterestimation module
 ---------------------------------------
 
 .. automodule:: basico.task_parameterestimation
-   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
@@ -67,7 +61,6 @@ basico.task\_steadystate module
 -------------------------------
 
 .. automodule:: basico.task_steadystate
-   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
@@ -76,7 +69,6 @@ basico.task\_timecourse module
 ------------------------------
 
 .. automodule:: basico.task_timecourse
-   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
@@ -85,7 +77,6 @@ basico.task\_scan module
 ------------------------------
 
 .. automodule:: basico.task_scan
-   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
@@ -95,7 +86,6 @@ basico.task\_optimization module
 ------------------------------
 
 .. automodule:: basico.task_optimization
-   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
@@ -104,7 +94,6 @@ basico.task\_sensitivities module
 ------------------------------
 
 .. automodule:: basico.task_sensitivities
-   :imported-members:
    :members:
    :undoc-members:
    :show-inheritance:
