### 2023-06-21 3ada13e
- add a function to simplify names
