### 2023-06-21 9660a2f
- experiment with running tasks collecting multiple outputs
