### 2023-06-29 f2fc65a
- fix warnings, return plot df
