+++ basico/model_info.py
@@ -157,8 +157,8 @@ def __status_to_string(status):
     return strings.get(status, 'fixed')
 
 
-def __xml_task_type_to_string(type):
-    return T.from_enum(type)
+def __xml_task_type_to_string(type_value):
+    return T.from_enum(type_value)
 
 
 def __task_name_to_xml_type(name):
@@ -398,7 +398,7 @@ def get_species(name=None, exact=False, **kwargs):
         if 'type' in kwargs and kwargs['type'] not in metab_data['type']:
             continue
 
-        if 'sbml_id' in kwargs and kwargs['sbml_id']  != metab_data['sbml_id']:
+        if 'sbml_id' in kwargs and kwargs['sbml_id'] != metab_data['sbml_id']:
             continue
 
         data.append(metab_data)
@@ -955,7 +955,7 @@ def get_default_plot_names(filter=None, **kwargs):
     if _default_plots:
         return _default_plots
 
-    _default_plots={}
+    _default_plots = {}
     for index in ids:
         item = COPASI.COutputAssistant.getItem(index)
         assert (isinstance(item, COPASI.CDefaultOutputDescription))
@@ -1062,7 +1062,8 @@ def add_report(name, **kwargs):
 
 
 def set_report_dict(spec, precision=None, separator=None, table=None,
-                    print_headers=True, header=None, body=None, footer=None, task=None, comment=None, add_separator=None, **kwargs):
+                    print_headers=True, header=None, body=None, footer=None, task=None,
+                    comment=None, add_separator=None, **kwargs):
     """Sets properties of the named report definition.
     
         Examples:
@@ -1113,8 +1114,8 @@ def set_report_dict(spec, precision=None, separator=None, table=None,
         :param comment: a documentation string for the report (can bei either string, or xhtml string)
         :type comment: Optional[str]
         
-        :param add_separator: an optional boolean flag, to automatically add seprators between header, body and footer entries
-               since this is not necessary for table entries. 
+        :param add_separator: an optional boolean flag, to automatically add seprators between header, body and
+            footer entries since this is not necessary for table entries.
         :type add_separator: Optional[bool]
 
         :param kwargs: optional arguments
@@ -1212,7 +1213,7 @@ def _set_report_vector(vec, list_of_cns, dm, separator=None):
             vec.append(COPASI.CRegisteredCommonName(separator))
 
     if separator:
-       vec.pop()
+        vec.pop()
 
 
 def wrap_copasi_string(text):
@@ -1222,6 +1223,7 @@ def wrap_copasi_string(text):
     :return: an escaped COPASI string, that can be used in reports
     """
     return 'String=' + COPASI.CCommonName.escape(text)
+
 
 def _replace_names_with_cns(expression, **kwargs):
     """ replaces all names in the given expression with cns
@@ -1820,7 +1822,6 @@ def remove_function(name, **kwargs):
         return
 
     key = fun.getKey()
-    fun = None
 
     db.removeFunction(key)
 
@@ -1948,7 +1949,8 @@ def add_event(name, trigger, assignments, **kwargs):
              * 'delay': the delay expression
              * 'priority': the priority expression
              * 'persistent': boolean indicating if the event is persistent
-             * 'delay_calculation': boolean indicating whether just the assignment is delayed, or the calculation as well
+             * 'delay_calculation': boolean indicating whether just the assignment is delayed, or the
+                                    calculation as well
              * 'fire_at_initial_time': boolean indicating if the event should fire at the initial time
 
         :return: the newly created event
[…]
