### 2023-06-29 a237ff0
- always compute criterion
