### 2023-06-29 6adcf1a
- add parameter to as_dict, so it would always return a list of elements if specified
