### 2023-06-29 0d2ffa7
- add parameter to as_dict, so it would always return a list of elements if specified
