### 2023-06-29 6c28580
- add basic plotting support
