### 2023-06-29 a50c5bd
- add basic plotting support
