### 2023-07-06 6ddf2cc
#34 : add compact parameter that was added in COPASI 4.36
