### 2023-07-06 21b45fb
#35 : filter error messages produced to display the important ones for simulations
This also updates the notebook, to ensure the correct error message `At least one particle number in the initial state is too big` is displayed.
