+++ tests/test_timecourse.py
@@ -24,7 +24,7 @@ class TestTimeCourse(unittest.TestCase):
         basico.set_reaction_parameters('(R1).k1', value=1)
         basico.set_reaction_parameters('(R2).k1', value=1)
         basico.add_event('E1', 'A <= 10', [('B', '50')])
-        result = basico.run_time_course(duration=10, method='stochastic', use_numbers=True)
+        result = basico.run_time_course(duration=10, method='directMethod', use_numbers=True)
         self.assertIsNotNone(result)
 
         # test that we can set hybrid methods
@@ -32,6 +32,15 @@ class TestTimeCourse(unittest.TestCase):
         self.assertIsNotNone(result)
         settings = basico.get_task_settings(basico.T.TIME_COURSE)
         self.assertEqual(settings['method']['name'], 'Hybrid (Runge-Kutta)')
+
+        # test that we cannot run the time course with too many particles
+        # and that the error message is correct:
+        basico.set_species('A', initial_particle_number=1e+24)
+        with self.assertLogs('basico', level='ERROR') as cm:
+            basico.run_time_course(duration=10, method='directMethod', use_numbers=True)
+            self.assertTrue(len(cm.output) > 0)
+            error_message = cm.output[0]
+            self.assertTrue('At least one particle number in the initial state is too big' in error_message)
 
     def test_stoch2(self):
         basico.new_model(name='Simple Model', substance_unit='1')
