### 2023-07-17 b532205
- add support for constraints in opt and fitproblems
requires newer python-copasi (4.41+) package.
