### 2023-07-17 3ebff23
- add support for constraints in opt and fitproblems
requires newer python-copasi (4.41+) package.
