### 2023-07-18 4eb0378
- run test for constraint only when supported
