### 2023-07-18 082e08f
- fix pytest warnings
