+++ paper.bib
@@ -117,3 +117,110 @@ available from http://www.sbml.org/Contact: sysbio-team@caltech.edu*To whom corr
  publisher = {USENIX Association},
  address = {Berkeley, CA, USA},
 } 
+
+@article{pyABC, doi = {10.21105/joss.04304}, url = {https://doi.org/10.21105/joss.04304}, year = {2022}, publisher = {The Open Journal}, volume = {7}, number = {74}, pages = {4304}, author = {Yannik Schälte and Emmanuel Klinger and Emad Alamoudi and Jan Hasenauer}, title = {pyABC: Efficient and robust easy-to-use approximate Bayesian computation}, journal = {Journal of Open Source Software} }
+
+@misc{sbmltoodejax,
+      title={SBMLtoODEjax: efficient simulation and optimization of ODE SBML models in JAX}, 
+      author={Mayalen Etcheverry and Michael Levin and Clément Moulin-Frier and Pierre-Yves Oudeyer},
+      year={2023},
+      eprint={2307.08452},
+      archivePrefix={arXiv},
+      primaryClass={q-bio.BM}
+	  doi = {10.48550/arXiv.2307.08452},
+	  ulr = {https://doi.org/10.48550/arXiv.2307.08452}
+}
+
+@ARTICLE{mendes2023reproducibility,
+  
+AUTHOR={Mendes, Pedro},   
+	 
+TITLE={Reproducibility and FAIR principles: the case of a segment polarity network model},      
+	
+JOURNAL={Frontiers in Cell and Developmental Biology},      
+	
+VOLUME={11},           
+	
+YEAR={2023},      
+	  
+URL={https://www.frontiersin.org/articles/10.3389/fcell.2023.1201673},       
+	
+DOI={10.3389/fcell.2023.1201673},      
+	
+ISSN={2296-634X},   
+   
+ABSTRACT={The issue of reproducibility of computational models and the related FAIR principles (findable, accessible, interoperable, and reusable) are examined in a specific test case. I analyze a computational model of the segment polarity network in Drosophila embryos published in 2000. Despite the high number of citations to this publication, 23 years later the model is barely accessible, and consequently not interoperable. Following the text of the original publication allowed successfully encoding the model for the open source software COPASI. Subsequently saving the model in the SBML format allowed it to be reused in other open source software packages. Submission of this SBML encoding of the model to the BioModels database enables its findability and accessibility. This demonstrates how the FAIR principles can be successfully enabled by using open source software, widely adopted standards, and public repositories, facilitating reproducibility and reuse of computational cell biology models that will outlive the specific software used.}
+}
+
+@software{the_pandas_development_team_2023_8092754,
+  author       = {The pandas development team},
+  title        = {pandas-dev/pandas: Pandas},
+  month        = jun,
+  year         = 2023,
+  publisher    = {Zenodo},
+  version      = {v2.0.3},
+  doi          = {10.5281/zenodo.8092754},
+  url          = {https://doi.org/10.5281/zenodo.8092754}
+}
+
+@software{libCombine,
+  author       = {Frank Bergmann and
+                  Matthias König and
+                  Jonathan Karr and
+                  Liam Keegan and
+                  Jitka Plesníková},
+  title        = {sbmlteam/libCombine: Release 0.2.20},
+  month        = dec,
+  year         = 2022,
+  publisher    = {Zenodo},
+  version      = {v0.2.20},
+  doi          = {10.5281/zenodo.7408732},
+  url          = {https://doi.org/10.5281/zenodo.7408732}
+}
+
+@software{libSEDML,
+  author       = {Frank Bergmann and
+                  Lucian Smith and
+                  David Nickerson and
+                  Alan Garny and
+                  Kyle Medley and
+                  Dmitry Marakasov},
[…]
