+++ paper.bib
@@ -126,9 +126,9 @@ available from http://www.sbml.org/Contact: sysbio-team@caltech.edu*To whom corr
       year={2023},
       eprint={2307.08452},
       archivePrefix={arXiv},
-      primaryClass={q-bio.BM}
-	  doi = {10.48550/arXiv.2307.08452},
-	  ulr = {https://doi.org/10.48550/arXiv.2307.08452}
+      primaryClass={q-bio.BM},
+	    doi = {10.48550/arXiv.2307.08452},
+	    ulr = {https://doi.org/10.48550/arXiv.2307.08452}
 }
 
 @ARTICLE{mendes2023reproducibility,
