### 2023-08-10 cf50c30
- change colab links for absolute references
