+++ paper.md
@@ -110,8 +110,8 @@ BioModels Database [@BioModels2015b] or JWS Online [@JWS].
 Of course the wrappers for the REST API to JWS Online or the BioModels Database can also be readily used by other Python packages to
 obtain the SBML models. This is doen for example by SBMLtoODEjax [@sbmltoodejax]
 
-Once a model is loaded all of `COPASI`'s analysis methods can be used. We started with 
-the implementation of time course simulations and steady state analysis. 
+Once a model is loaded all of `COPASI`'s analysis methods can be used. Running simulations are a core feature of COPASI, so we started 
+We started `BASICO` with implementing time course simulations and steady state analysis. 
 
 ```python
 
@@ -121,7 +121,7 @@ the implementation of time course simulations and steady state analysis.
 ```
 
 Here, `run_time_course` returns a `pandas.DataFrame` [@pandas] with the results of the simulation. We 
-quickly received requests to add further analysis methods, and so we added parameter estimation,
+quickly received feature requests to add further analysis methods, and so we added parameter estimation,
 optimizations, sensitivity analysis and parameter scans. 
 
 Most recently we added the automation of profile likelihood calculations, that for a given
