+++ paper.bib
@@ -152,7 +152,7 @@ ISSN={2296-634X},
 ABSTRACT={The issue of reproducibility of computational models and the related FAIR principles (findable, accessible, interoperable, and reusable) are examined in a specific test case. I analyze a computational model of the segment polarity network in Drosophila embryos published in 2000. Despite the high number of citations to this publication, 23 years later the model is barely accessible, and consequently not interoperable. Following the text of the original publication allowed successfully encoding the model for the open source software COPASI. Subsequently saving the model in the SBML format allowed it to be reused in other open source software packages. Submission of this SBML encoding of the model to the BioModels database enables its findability and accessibility. This demonstrates how the FAIR principles can be successfully enabled by using open source software, widely adopted standards, and public repositories, facilitating reproducibility and reuse of computational cell biology models that will outlive the specific software used.}
 }
 
-@software{the_pandas_development_team_2023_8092754,
+@software{pandas,
   author       = {The pandas development team},
   title        = {pandas-dev/pandas: Pandas},
   month        = jun,
