+++ paper.bib
@@ -153,7 +153,7 @@ ABSTRACT={The issue of reproducibility of computational models and the related F
 }
 
 @software{pandas,
-  author       = {The pandas development team},
+  author       = {{pandas development team}},
   title        = {pandas-dev/pandas: Pandas},
   month        = jun,
   year         = 2023,
