### 2023-08-13 #5553 @jguhlin
Hello. My review is short. Installing the software on different operating systems was no problem; the examples from the repository ran with no issues, and the documentation is easy to read and follow. I did not test the jupyter notebooks but only the command-line examples  Especially with the comments and changes above, I believe only minor textual edits are required.

Minor text changes:
Pg 2 line 70. "Running simulations are a core feature of COPASI, so we started We started BASICO with implementing time course
simulations and steady state analysis"

"We started" is there twice.

Pg 2 line 78. "This is another example, that would have been quite cumbersome to do without basico"
BASICO should be capitalized.

Pg 2. Line 51. "As scripting module BASICO lends itself for constructing large networks, as is for example done51
in the reproducibility study in (Mendes, 2023)."
Remove parentheses around citation (but otherwise keep as citation), if possible.

Very minor change:
"At the same time its broad range of tasks, makes it attractive for more advanced computational biologists"
to:
"At the same time, its broad range of tasks makes it attractive for more advanced computational biologists"

Just moving the comma.

Please let me know if any of those are difficult to achieve. I have no other notes, well done on the library!
