### 2023-08-14 #5553 @fbergmann
Thank you very much for your review @jguhlin, I've addressed all the issues as follows:

> Pg 2 line 70. "Running simulations are a core feature of COPASI, so we started We started BASICO with implementing time course
simulations and steady state analysis"
> "We started" is there twice.

thank you for catching that

> Pg 2 line 78. "This is another example, that would have been quite cumbersome to do without basico"
BASICO should be capitalized.

Indeed!

> Pg 2. Line 51. "As scripting module BASICO lends itself for constructing large networks, as is for example done
in the reproducibility study in (Mendes, 2023)."
Remove parentheses around citation (but otherwise keep as citation), if possible.

Indeed this looks much better without the parentheses using the other reference style this will appear as "... Mendes (2023)."

> Very minor change:
"At the same time its broad range of tasks, makes it attractive for more advanced computational biologists"
to:
"At the same time, its broad range of tasks makes it attractive for more advanced computational biologists"

> Just moving the comma.

Thank you again!
