### 2023-09-24 #5553 @darogan
My review @fbergmann @marcosvital for BASICO

BASICO provides a python interface to COPASI enabling the more automated and high-throughput analyses. This is a very welcome endeavour and should open up COPASI to a wider audience.

Paper:
I have some minor comments to the text
1. ln25 change mathematical backgrounds to underlying mathematical concepts
2. Add a brief description of SBML (what is stands for etc)
3. ln47 readily *be* integrated
4. ln50 done -> achieved
5. ln51 as *a* scripting
6. ln60 SED-ML would aid the reader to define what this stands for
7. ln73 test it a little informal. Suggest updating to something on the lines of "Feature requests from the community led us to add further ..."
8. ln76 Suggest changing to a more formal style e.g. "The most recent additions include ..."

Code:
I was able to run though several examples using Jupyter with ease. Author should be commended on a simple installation and excellent documentation.
I didn't try the google colab options, so can't comment on ease of use here
