### 2023-10-05 #5553 @marcosvital
@fbergmann it seems that we are almos ready to finish. Please take a look at the last version of the manuscript, and let me know if everything is ok.

Also, if you already didn't do this: you will need to archive the last release of the package (on Zenodo, figshare, or other of your choice) and share the archive DOI here.
