### 2023-10-06 #5553 @fbergmann
@marcosvital  I'm happy with the final version. And I do have the automatic zenodo integration going, that will create a release package for each release. The last one being currently <https://zenodo.org/record/8189386>. Please let me know if I can do anything else to expedite.
