### 2023-10-15 #5553 @fbergmann
@Kevin-Mattheus-Moerman thank you for looking things over. I've changed the zenodo archive  title in [10.5281/zenodo.8189386](https://doi.org/10.5281/zenodo.8189386)  to read `BASICO: A simplified Python interface to COPASI`. Just as it is in the final version of the paper.
