### 2023-10-16 f85310d
update citation with joss reference
