+++ CITATION.cff
@@ -1,9 +1,26 @@
-cff-version: 1.1.0
-message: "If you use this software, please cite it as below."
+cff-version: "1.2.0"
 authors:
+- family-names: Bergmann
+  given-names: Frank T.
+  orcid: "https://orcid.org/0000-0001-5553-4702"
+doi: 10.5281/zenodo.8189386
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
   - family-names: Bergmann
     given-names: Frank T.
-    orcid: https://orcid.org/0000-0001-5553-4702
-title: "copasi/basico: Release 0.48"
-version: v0.48
-date-released: 2023-03-29
+    orcid: "https://orcid.org/0000-0001-5553-4702"
+  date-published: 2023-10-15
+  doi: 10.21105/joss.05553
+  issn: 2475-9066
+  issue: 90
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 5553
+  title: "BASICO: A simplified Python interface to COPASI"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.05553"
+  volume: 8
+title: "BASICO: A simplified Python interface to COPASI"
